fix: rename project to gpustack-helper
