+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "gpustackhelper"
+name = "gpustack-helper"
 version = "0.1.0"
 description = "The GPUStack Installer is a helper tool for installing and configuring GPUStack."
 authors = ["GPUStack Authors <contact@gpustack.ai>"]
@@ -22,4 +22,4 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-helper = "gpustackhelper.main:main"
+helper = "gpustack_helper.main:main"
